examples/scaffolder templates/cookiecutter-circleci/requirements.txt
pytest-cookies==0.3.0
cookiecutter==1.6.0
